added documentation for plugins and did some clean up Repaired console input
