+++ papi/plugin/visual/Console/Console.py
@@ -30,7 +30,6 @@ Contributors:
 
 
 from PyQt5 import QtCore
-
 from PyQt5.QtWidgets import QWidget
 from PyQt5 import QtWidgets
 from PyQt5 import QtGui
@@ -38,12 +37,50 @@ from papi.plugin.base_classes.vip_base import vip_base
 from papi.data.DParameter import DParameter
 from papi.data.DPlugin import DBlock
 
-from papi.plugin.visual.Console.CmdInput import CmdInput
+
+
+
+class CmdInput(QtWidgets.QLineEdit):
+
+    sigExecuteCmd = QtCore.pyqtSignal(object)
+
+    def __init__(self, parent):
+        QtWidgets.QLineEdit.__init__(self, parent)
+        self.history = [""]
+        self.ptr = 0
+
+    def keyPressEvent(self, ev):
+        if ev.key() == QtCore.Qt.Key_Up and self.ptr < len(self.history) - 1:
+            self.setHistory(self.ptr+1)
+            ev.accept()
+            return
+        elif ev.key() ==  QtCore.Qt.Key_Down and self.ptr > 0:
+            self.setHistory(self.ptr-1)
+            ev.accept()
+            return
+        elif ev.key() == QtCore.Qt.Key_Return:
+            self.execCmd()
+        else:
+            QtWidgets.QLineEdit.keyPressEvent(self, ev)
+            self.history[0] = (self.text())
+
+    def execCmd(self):
+        cmd = (self.text())
+        if len(self.history) == 1 or cmd != self.history[1]:
+            self.history.insert(1, cmd)
+
+        self.history[0] = ""
+        self.setHistory(0)
+        self.sigExecuteCmd.emit(cmd)
+
+    def setHistory(self, num):
+        self.ptr = num
+        self.setText(self.history[self.ptr])
+
+
 
 
 class Console(vip_base):
-
-
     def cb_initialize_plugin(self):
 
         # ---------------------------
